reorganize generic objectives
