+++ desc/objectives/__init__.py
@@ -9,7 +9,7 @@ from ._equilibrium import (
     HelicalForceBalance,
     RadialForceBalance,
 )
-from ._generic import GenericObjective, ObjectiveFromUser
+from ._generic import GenericObjective, LinearObjectiveFromUser, ObjectiveFromUser
 from ._geometry import (
     AspectRatio,
     BScaleLength,
@@ -62,6 +62,5 @@ from .linear_objectives import (
     FixSumModesR,
     FixSumModesZ,
     FixThetaSFL,
-    LinearObjectiveFromUser,
 )
 from .objective_funs import ObjectiveFunction
