+++ desc/objectives/_generic.py
@@ -9,7 +9,264 @@ from desc.compute import data_index
 from desc.compute.utils import get_profiles, get_transforms
 from desc.grid import QuadratureGrid
 
+from .linear_objectives import _FixedObjective
 from .objective_funs import _Objective
+
+
+class GenericObjective(_Objective):
+    """A generic objective that can compute any quantity from the `data_index`.
+
+    Parameters
+    ----------
+    f : str
+        Name of the quantity to compute.
+    eq : Equilibrium
+        Equilibrium that will be optimized to satisfy the Objective.
+    target : {float, ndarray}, optional
+        Target value(s) of the objective. Only used if bounds is None.
+        Must be broadcastable to Objective.dim_f.
+    bounds : tuple of {float, ndarray}, optional
+        Lower and upper bounds on the objective. Overrides target.
+        Both bounds must be broadcastable to to Objective.dim_f
+    weight : {float, ndarray}, optional
+        Weighting to apply to the Objective, relative to other Objectives.
+        Must be broadcastable to to Objective.dim_f
+    normalize : bool, optional
+        Whether to compute the error in physical units or non-dimensionalize.
+        Has no effect for this objective
+    normalize_target : bool, optional
+        Whether target and bounds should be normalized before comparing to computed
+        values. If `normalize` is `True` and the target is in physical units,
+        this should also be set to True. Note: Has no effect on this objective.
+    loss_function : {None, 'mean', 'min', 'max'}, optional
+        Loss function to apply to the objective values once computed. This loss function
+        is called on the raw compute value, before any shifting, scaling, or
+        normalization.
+    deriv_mode : {"auto", "fwd", "rev"}
+        Specify how to compute jacobian matrix, either forward mode or reverse mode AD.
+        "auto" selects forward or reverse mode based on the size of the input and output
+        of the objective. Has no effect on self.grad or self.hess which always use
+        reverse mode and forward over reverse mode respectively.
+    grid : Grid, optional
+        Collocation grid containing the nodes to evaluate at.
+    name : str, optional
+        Name of the objective function.
+
+    """
+
+    _print_value_fmt = "GenericObjective value: {:10.3e} "
+
+    def __init__(
+        self,
+        f,
+        eq,
+        target=None,
+        bounds=None,
+        weight=1,
+        normalize=True,
+        normalize_target=True,
+        loss_function=None,
+        deriv_mode="auto",
+        grid=None,
+        name="generic",
+    ):
+        if target is None and bounds is None:
+            target = 0
+        self.f = f
+        self._grid = grid
+        super().__init__(
+            things=eq,
+            target=target,
+            bounds=bounds,
+            weight=weight,
+            normalize=normalize,
+            normalize_target=normalize_target,
+            loss_function=loss_function,
+            deriv_mode=deriv_mode,
+            name=name,
+        )
+        self._scalar = not bool(
+            data_index["desc.equilibrium.equilibrium.Equilibrium"][self.f]["dim"]
+        )
+        self._coordinates = data_index["desc.equilibrium.equilibrium.Equilibrium"][
+            self.f
+        ]["coordinates"]
+        self._units = (
+            "("
+            + data_index["desc.equilibrium.equilibrium.Equilibrium"][self.f]["units"]
+            + ")"
+        )
+
+    def build(self, use_jit=True, verbose=1):
+        """Build constant arrays.
+
+        Parameters
+        ----------
+        use_jit : bool, optional
+            Whether to just-in-time compile the objective and derivatives.
+        verbose : int, optional
+            Level of output.
+
+        """
+        eq = self.things[0]
+        if self._grid is None:
+            grid = QuadratureGrid(eq.L_grid, eq.M_grid, eq.N_grid, eq.NFP)
+        else:
+            grid = self._grid
+
+        p = "desc.equilibrium.equilibrium.Equilibrium"
+        if data_index[p][self.f]["dim"] == 0:
+            self._dim_f = 1
+        elif data_index[p][self.f]["coordinates"] == "r":
+            self._dim_f = grid.num_rho
+        else:
+            self._dim_f = grid.num_nodes * data_index[p][self.f]["dim"]
+        profiles = get_profiles(self.f, obj=eq, grid=grid)
+        transforms = get_transforms(self.f, obj=eq, grid=grid)
+        self._constants = {
+            "transforms": transforms,
+            "profiles": profiles,
+        }
+
+        super().build(use_jit=use_jit, verbose=verbose)
+
+    def compute(self, params, constants=None):
+        """Compute the quantity.
+
+        Parameters
+        ----------
+        params : dict
+            Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
+        constants : dict
+            Dictionary of constant data, eg transforms, profiles etc. Defaults to
+            self.constants
+
+        Returns
+        -------
+        f : ndarray
+            Computed quantity.
+
+        """
+        if constants is None:
+            constants = self.constants
+        data = compute_fun(
+            "desc.equilibrium.equilibrium.Equilibrium",
+            self.f,
+            params=params,
+            transforms=constants["transforms"],
+            profiles=constants["profiles"],
+        )
+        f = data[self.f]
+        if self._coordinates == "r":
+            f = constants["transforms"]["grid"].compress(f, surface_label="rho")
+        return f.flatten(order="F")  # so that default quad weights line up correctly.
+
+
+class LinearObjectiveFromUser(_FixedObjective):
+    """Wrap a user defined linear objective function.
+
+    The user supplied function should take one argument, ``params``, which is a
+    dictionary of parameters of an Optimizable "thing".
+
+    The function should be JAX traceable and differentiable, and should return a single
+    JAX array. The source code of the function must be visible to the ``inspect`` module
+    for parsing.
+
+    Parameters
+    ----------
+    fun : callable
+        Custom objective function.
+    thing : Optimizable
+        Object whose degrees of freedom are being constrained.
+    target : dict of {float, ndarray}, optional
+        Target value(s) of the objective. Only used if bounds is None.
+        Should have the same tree structure as thing.params. Defaults to things.params.
+    bounds : tuple of dict {float, ndarray}, optional
+        Lower and upper bounds on the objective. Overrides target.
+        Should have the same tree structure as thing.params.
+    weight : dict of {float, ndarray}, optional
+        Weighting to apply to the Objective, relative to other Objectives.
+        Should be a scalar or have the same tree structure as thing.params.
+    normalize : bool, optional
+        Whether to compute the error in physical units or non-dimensionalize.
+        Has no effect for this objective.
+    normalize_target : bool, optional
+        Whether target and bounds should be normalized before comparing to computed
+        values. If `normalize` is `True` and the target is in physical units,
+        this should also be set to True. Has no effect for this objective.
+    name : str, optional
+        Name of the objective function.
+
+    """
+
+    _scalar = False
+    _linear = True
+    _fixed = True
+    _units = "(Unknown)"
+    _print_value_fmt = "Custom linear Objective value: {:10.3e}"
+
+    def __init__(
+        self,
+        fun,
+        thing,
+        target=None,
+        bounds=None,
+        weight=1,
+        normalize=False,
+        normalize_target=False,
+        name="custom linear",
+    ):
+        if target is None and bounds is None:
+            target = 0
+        self._fun = fun
+        super().__init__(
+            things=thing,
+            target=target,
+            bounds=bounds,
+            weight=weight,
+            normalize=normalize,
+            normalize_target=normalize_target,
+            name=name,
+        )
+
+    def build(self, use_jit=False, verbose=1):
+        """Build constant arrays.
+
+        Parameters
+        ----------
+        use_jit : bool, optional
+            Whether to just-in-time compile the objective and derivatives.
+        verbose : int, optional
+            Level of output.
+
+        """
+        thing = self.things[0]
+
+        import jax
+
+        self._dim_f = self._dim_f = jax.eval_shape(self._fun, thing.params_dict).size
+
+        super().build(use_jit=use_jit, verbose=verbose)
+
+    def compute(self, params, constants=None):
+        """Compute fixed degree of freedom errors.
+
+        Parameters
+        ----------
+        params : dict
+            Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
+        constants : dict
+            Dictionary of constant data, eg transforms, profiles etc. Defaults to
+            self.constants
+
+        Returns
+        -------
+        f : ndarray
+            Fixed degree of freedom errors.
+
+        """
+        f = self._fun(params)
+        return f
 
 
 class ObjectiveFromUser(_Objective):
@@ -195,152 +452,3 @@ class ObjectiveFromUser(_Objective):
         )
         f = self._fun_wrapped(data)
         return f
-
-
-class GenericObjective(_Objective):
-    """A generic objective that can compute any quantity from the `data_index`.
-
-    Parameters
-    ----------
-    f : str
-        Name of the quantity to compute.
-    eq : Equilibrium
-        Equilibrium that will be optimized to satisfy the Objective.
-    target : {float, ndarray}, optional
-        Target value(s) of the objective. Only used if bounds is None.
-        Must be broadcastable to Objective.dim_f.
-    bounds : tuple of {float, ndarray}, optional
-        Lower and upper bounds on the objective. Overrides target.
-        Both bounds must be broadcastable to to Objective.dim_f
-    weight : {float, ndarray}, optional
-        Weighting to apply to the Objective, relative to other Objectives.
-        Must be broadcastable to to Objective.dim_f
-    normalize : bool, optional
-        Whether to compute the error in physical units or non-dimensionalize.
-        Has no effect for this objective
-    normalize_target : bool, optional
-        Whether target and bounds should be normalized before comparing to computed
-        values. If `normalize` is `True` and the target is in physical units,
-        this should also be set to True. Note: Has no effect on this objective.
-    loss_function : {None, 'mean', 'min', 'max'}, optional
-        Loss function to apply to the objective values once computed. This loss function
-        is called on the raw compute value, before any shifting, scaling, or
-        normalization.
-    deriv_mode : {"auto", "fwd", "rev"}
-        Specify how to compute jacobian matrix, either forward mode or reverse mode AD.
-        "auto" selects forward or reverse mode based on the size of the input and output
-        of the objective. Has no effect on self.grad or self.hess which always use
-        reverse mode and forward over reverse mode respectively.
-    grid : Grid, optional
-        Collocation grid containing the nodes to evaluate at.
-    name : str, optional
-        Name of the objective function.
-
-    """
-
-    _print_value_fmt = "GenericObjective value: {:10.3e} "
-
-    def __init__(
-        self,
-        f,
-        eq,
-        target=None,
-        bounds=None,
-        weight=1,
-        normalize=True,
-        normalize_target=True,
-        loss_function=None,
-        deriv_mode="auto",
-        grid=None,
-        name="generic",
-    ):
-        if target is None and bounds is None:
-            target = 0
-        self.f = f
-        self._grid = grid
-        super().__init__(
-            things=eq,
-            target=target,
-            bounds=bounds,
-            weight=weight,
-            normalize=normalize,
-            normalize_target=normalize_target,
-            loss_function=loss_function,
-            deriv_mode=deriv_mode,
-            name=name,
-        )
-        self._scalar = not bool(
-            data_index["desc.equilibrium.equilibrium.Equilibrium"][self.f]["dim"]
-        )
-        self._coordinates = data_index["desc.equilibrium.equilibrium.Equilibrium"][
-            self.f
-        ]["coordinates"]
-        self._units = (
-            "("
-            + data_index["desc.equilibrium.equilibrium.Equilibrium"][self.f]["units"]
-            + ")"
-        )
-
-    def build(self, use_jit=True, verbose=1):
-        """Build constant arrays.
-
-        Parameters
-        ----------
-        use_jit : bool, optional
-            Whether to just-in-time compile the objective and derivatives.
-        verbose : int, optional
-            Level of output.
-
-        """
-        eq = self.things[0]
-        if self._grid is None:
-            grid = QuadratureGrid(eq.L_grid, eq.M_grid, eq.N_grid, eq.NFP)
-        else:
-            grid = self._grid
-
-        p = "desc.equilibrium.equilibrium.Equilibrium"
-        if data_index[p][self.f]["dim"] == 0:
-            self._dim_f = 1
-        elif data_index[p][self.f]["coordinates"] == "r":
-            self._dim_f = grid.num_rho
-        else:
-            self._dim_f = grid.num_nodes * data_index[p][self.f]["dim"]
-        profiles = get_profiles(self.f, obj=eq, grid=grid)
-        transforms = get_transforms(self.f, obj=eq, grid=grid)
-        self._constants = {
-            "transforms": transforms,
-            "profiles": profiles,
-        }
-
-        super().build(use_jit=use_jit, verbose=verbose)
-
-    def compute(self, params, constants=None):
-        """Compute the quantity.
-
-        Parameters
-        ----------
-        params : dict
-            Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
-        constants : dict
-            Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
-
-        Returns
-        -------
-        f : ndarray
-            Computed quantity.
-
-        """
-        if constants is None:
-            constants = self.constants
-        data = compute_fun(
-            "desc.equilibrium.equilibrium.Equilibrium",
-            self.f,
-            params=params,
-            transforms=constants["transforms"],
-            profiles=constants["profiles"],
-        )
-        f = data[self.f]
-        if self._coordinates == "r":
-            f = constants["transforms"]["grid"].compress(f, surface_label="rho")
-        return f.flatten(order="F")  # so that default quad weights line up correctly.
