+++ desc/objectives/linear_objectives.py
@@ -209,106 +209,6 @@ class FixParameter(_FixedObjective):
         )
 
 
-class LinearObjectiveFromUser(_FixedObjective):
-    """TODO: description.
-
-    Parameters
-    ----------
-    fun : callable
-        Custom objective function.
-    thing : Optimizable
-        Object whose degrees of freedom are being fixed.
-    target : dict of {float, ndarray}, optional
-        Target value(s) of the objective. Only used if bounds is None.
-        Should have the same tree structure as thing.params. Defaults to things.params.
-    bounds : tuple of dict {float, ndarray}, optional
-        Lower and upper bounds on the objective. Overrides target.
-        Should have the same tree structure as thing.params.
-    weight : dict of {float, ndarray}, optional
-        Weighting to apply to the Objective, relative to other Objectives.
-        Should be a scalar or have the same tree structure as thing.params.
-    normalize : bool, optional
-        Whether to compute the error in physical units or non-dimensionalize.
-        Has no effect for this objective.
-    normalize_target : bool, optional
-        Whether target and bounds should be normalized before comparing to computed
-        values. If `normalize` is `True` and the target is in physical units,
-        this should also be set to True. Has no effect for this objective.
-    name : str, optional
-        Name of the objective function.
-
-    """
-
-    _scalar = False
-    _linear = True
-    _fixed = True
-    _units = "(Unknown)"
-    _print_value_fmt = "Custom linear Objective value: {:10.3e}"
-
-    def __init__(
-        self,
-        fun,
-        thing,
-        target=None,
-        bounds=None,
-        weight=1,
-        normalize=False,
-        normalize_target=False,
-        name="custom linear",
-    ):
-        if target is None and bounds is None:
-            target = 0
-        self._fun = fun
-        super().__init__(
-            things=thing,
-            target=target,
-            bounds=bounds,
-            weight=weight,
-            normalize=normalize,
-            normalize_target=normalize_target,
-            name=name,
-        )
-
-    def build(self, use_jit=False, verbose=1):
-        """Build constant arrays.
-
-        Parameters
-        ----------
-        use_jit : bool, optional
-            Whether to just-in-time compile the objective and derivatives.
-        verbose : int, optional
-            Level of output.
-
-        """
-        thing = self.things[0]
-
-        import jax
-
-        self._dim_f = self._dim_f = jax.eval_shape(self._fun, thing.params_dict).size
-
-        super().build(use_jit=use_jit, verbose=verbose)
-
-    def compute(self, params, constants=None):
-        """Compute fixed degree of freedom errors.
-
-        Parameters
-        ----------
-        params : dict
-            Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
-        constants : dict
-            Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
-
-        Returns
-        -------
-        f : ndarray
-            Fixed degree of freedom errors.
-
-        """
-        f = self._fun(params)
-        return f
-
-
 class BoundaryRSelfConsistency(_Objective):
     """Ensure that the boundary and interior surfaces are self-consistent.
